Jpeg, and also send it on every validation
