+++ src/privacy_review_assistant_group.py
@@ -1,12 +1,10 @@
 from autogen import AssistantAgent, GroupChat, GroupChatManager, Agent
-from autogen.agentchat.contrib.capabilities.agent_capability import AgentCapability
-from autogen.agentchat.contrib.img_utils import pil_to_data_uri
 from botbuilder.core import TurnContext
-from PIL import Image
 
 from state import AppTurnState
 from rag_agents import setup_rag_assistant
-from threat_model_reviewer_group import ThreatModelReviewerGroup, ThreatModelImageVisualizer
+from threat_model_reviewer_group import ThreatModelReviewerGroup
+from threat_model_visualizer import ThreatModelImageVisualizerCapability
 
 class PrivacyReviewAssistantGroup:
     def __init__(self, llm_config):
@@ -66,32 +64,3 @@ class PrivacyReviewAssistantGroup:
         visualizer_capability = ThreatModelImageVisualizerCapability(state=state)
         visualizer_capability.add_to_agent(visualizer_assistant)
         return visualizer_assistant
-    
-class ThreatModelImageVisualizerCapability(AgentCapability, ThreatModelImageVisualizer):
-    def __init__(self, state: AppTurnState):
-        super().__init__()
-        super(AgentCapability, self).__init__(state)
-    
-    def add_to_agent(self, agent: AssistantAgent):
-        agent.register_reply([Agent, None], self._reply_with_image, remove_other_reply_funcs=True)
-        
-    def _reply_with_image(self, self2, messages, sender, config):
-        if self.img is None:
-            self.extract_image_from_state()
-            if self.img:
-                self._convert_to_jpeg_if_needed(self.img)
-            
-        if self.img:
-            return [True, self._convert_image_to_data_uri(self.img)]
-        else:
-            return [True, "No threat model available"]
-        
-    def _convert_to_jpeg_if_needed(self, image: Image.Image):
-        if image.mode != "RGB":
-            new_image = Image.new("RGBA", image.size, "WHITE") # Create a white rgba background
-            jpeg_img = Image.alpha_composite(new_image, image)
-            self.img = jpeg_img
-        return
-        
-    def _convert_image_to_data_uri(self, image: Image.Image):
-        return pil_to_data_uri(image)
